\name{plotZone}
\alias{plotZone}
\title{Plots a 2D tecplot structured zone.}
\usage{
  plotZone(z, add = FALSE, add.text = TRUE, cex = 1,
    outline = FALSE)
}
\arguments{
  \item{z}{Tecplot zone.}

  \item{add}{Add to an existing plot?}

  \item{add.text}{Plot the indices of each cell?}

  \item{cex}{Text font size multiplier.}

  \item{outline}{Draw the outline of the zone?.}
}
\description{
  Debuggin utilitity that plots a 2D tecplot structured
  zone and identifies each cell in the grid.
}

