\name{pnpoly}
\alias{pnpoly}
\title{Determine if point is inside a polygon.}
\usage{
  pnpoly(xv, yv, x, y)
}
\arguments{
  \item{xv}{X coordinates of the polygon's vertices.}

  \item{yv}{Y coordinates of the polygon's vertices.}

  \item{x}{X coordinate of the point.}

  \item{y}{Y coordinate of the point.}
}
\value{
  TRUE if the point is inside the polygon.
}
\description{
  This function determines whether a point is inside a
  polygon characterized by its vertices.
}

