% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecplot.R
\name{readTecplot}
\alias{readTecplot}
\title{Reads a tecplot structured grid file .}
\usage{
readTecplot(fname)
}
\arguments{
\item{fname}{Tecplot file name.}
}
\value{
List of zone data.
}
\description{
Reads a tecplot structured grid file with one or more zones.
}
