\name{transfTec}
\alias{transfTec}
\title{Returns a list of variables from a tecplot zone.}
\usage{
  transfTec(d)
}
\arguments{
  \item{d}{tecplot structured zone.}
}
\value{
  List of variable arrays.
}
\description{
  This function returns a list of arrays of each variable
  in a tecplot zone.
}

