\name{writeTecplot}
\alias{writeTecplot}
\title{Write a tecplot structured mesh to a file.}
\usage{
  writeTecplot(fname, tec, tectitle = NULL)
}
\arguments{
  \item{fname}{Name of the file to be created.}

  \item{tec}{List of zone data.}

  \item{tectitle}{Title of the tecplot file (optional).}
}
\description{
  Creates a tecplot structured mesh file.
}

