\name{zoneCoords}
\alias{zoneCoords}
\title{Returns zone coordinate array.}
\usage{
  zoneCoords(z)
}
\arguments{
  \item{z}{Tecplot zone.}
}
\value{
  List containing the x and y coordinates the the zone
  array.

  List of zone coordinates in array format.
}
\description{
  This function returns the coordinate arrays of a zone.
}

