\name{readTecFiles}
\alias{readTecFiles}
\title{Reads a list of tecplot files.}
\usage{
  readTecFiles(files)
}
\arguments{
  \item{files}{File names of structured meshes with same
  structure.}
}
\value{
  List with all the data loaded.
}
\description{
  This function loads a list of files which have the same
  structure. and returns a data structure with all the
  data.
}

